refactor: changed schemas import on SignInForm and SignUpForm to @schemas; Changed auth and register page names to sign-in and sign-up; Moved sign-in and sign-up to the root pages folder; Renamed styles to authStyles
