+++ src/pages/authStyles.ts
